+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGM98JAJQ8HF93976VWPZJ4F"
+        "01KGMDW3FTZDJ5RFXMRJ2GSFQP"
     )
     r"""The unique identifier of the data source"""